refactor(cli): split todo.go below 1000 lines
Move each todo cobra command into a sibling file. Keep
command wiring in todo.go. No behavior change.

Related: #652
